CSS Update
Finally fixed the sticky nav. Lots of changes to the CSS of the Main. Commented it all out so I can edit the HTML (add all the text I want), then will look to see if CSS still applicable and if so, comment back in.
Also, file names have changed to represent what I actually want to include on website and make it a more logical layout to navigate as a user